backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
PyMuPDF==1.25.3
EbookLib==0.18
mobi==0.3.3
httpx==0.28.1
python-multipart==0.0.20
Pillow==11.1.0
openai==1.61.0
pydantic-settings==2.7.1
beautifulsoup4==4.12.3
isbnlib==3.10.14
Wikipedia-API==0.11.0
setuptools>=70.0.0
